Fix deadlock with logcat service
Both logcat service and logcat listener hold their
respective lock around calls to each other. That
is a recipe for deadlocks.

This change ensures that the logcat listener holds its
internal lock only for internal only operations (read/write
to internal map) and moves the calls to the logcat service
outside the internal lock.

The locgat service should be fixed to do the same, but that
is a more complex endeavor, which may require a refactoring of
most of the code.

Bug: 67172466
Test: Existing tests
Change-Id: I259babe8c6f1495c93b3904b1964d7e7d52d1ba7