Handle sync errors after project is initialised.
Error handlers may access PSI structures (e.g., build files) and those
may throw if the project is not yet initialised.

Bug: b/66911744
Test: Added testHandleErrorWhenProjectNotInitialized for new sync. For
old sync, AndroidGradleTestCase is used for all error handlers, which
impacts the normal StartupManager workflow and adding a selector which
would invoke StartupManager depending on the current test case would be
an overkill. So for old sync a simple branching is implemented instead.
Change-Id: Iaf8283f1f9a7a91eee33f3c81d2e5487053d8efe