Switches to the preferred process only once
Previously, once a preferred process is set, the profiler
always switches back to it whenever it appears in the process
list, even if the user was intentionally profiling a different
app. The new behavior sets the preferred process when the user
performs the profiler action, and the profiler only switches
to the preferred process only once.

Bug: 67415158
Bug: 67123981
Test: updated unit test
Change-Id: I998c477093e3d2e98365147747ce4e8eaecd151f