Ask the user before adding a dependency.
This functionality broke with this change: I26c4fc128
That change fixes the case where we are dropping a cast item
on a menu by adding the dependencies in a transaction but
ask the user about the dependencies first.

Unfortunately the change broke the general case.

In this CL NlModel.addComponents will call NlModel.canAddComponents
before a write transaction is initiated.
Callers that must wrap NlModel.addComponents in their own write
transaction must call NlModel.canAddComponents before starting the
write transaction and call NlModel.addDependencies before calling
NlModel.addComponents.

A check that this is not violated is added as an assert in
NlModelHelper.canAddComponents. If this should happen in a deployed
version of AS the user will simply not be asked.

Bug: b/67464018
Test: Updated. Checks in ui-test.
Change-Id: Ia143a80d4ef791dd93c5d8f9ea70135f44b9f539