[New Sync] Setup ProjectData DataNode to display task view.
Creates DataNode for ProjectData, ModuleData and TASK,
then set the ProjectData as ExternalProjectInfo in
ProjectDataManager, which is used to feed task view in
Gradle tools window.

Add unit test to verify that ExternalProjectInfo is linked
correctly, and ProjectNode is not empty.
This CL doesn't contain the test that verifies contents of
task view. It will be in integration test in a separate cl.

Bug: 67739139
Test: ProjectDataNodeSetupTest.
Change-Id: Id3ac24d34af77902896ae5d8a0b1100c502399b1