Fix stack overflow problem in Slice Editor.
Bug: b/67711572

The table cell editor used for editing resource values is
using the same editor instance for all resources.

The stack overflow happened because:
- A <tab> key is causing focus to be transferred.
- We want the next resource value to receive focus next.
- In order to show the editor it must be added to the table.
- When adding the editor to the table, the editor must first
  be removed from its current parent (which is the table).
- The removeNotify logic in swing determines that focus must
  be transferred (which essentially is causing the loop).

Most of the above logic happens deep inside Swing. I found a
workaround by encapsulating the editor in a JPanel. That causes
the AutoFocusTransfer logic in Container.java to be disabled for
the editor.

After the stack overflow problem was fixed, there was a focus
problem left. On focus loss we would be closing the editor by
treating it as if the user typed <enter>. It turned out that when
tabbing from value to value the editor close was processed too
late, and it would cause the newly editor to close instead of
the editor that was loosing focus.
The fix was simply to remove the <enter> processing since the
table would also close the editor when focus is lost.

Test: Added a ui-test in a followup CL such that it is easy to
      identify the fixes from the test changes.

Change-Id: I1a69d44cf1f1c3637304b5a88e6381b8c20ef110