Remove obsolete test
We used to have a IDE-specific implementation of the API detector,
called LintIdeApiDetector, with its own unit test, LintIdeApiDetectorTest.
This custom version of the API detector was deleted in 2.3; there is
a single API Detector in lint now. When I merged the two detectors
I migrated all the IDE-side unit tests up to lint. However, I accidentally
left these tests around in the IDE plugin; they no longer need to be here.

Test: This deletes a test
Change-Id: I5db6716b5d45d0d2bff2686cbf14fc70e4636e2a