Use sync start time in GradleSyncState#isSyncNeeded
Currently, GradleSyncState#isSyncNeeded determines whether or not a
sync is necessary by checking to see if any Gradle files have been
modified since the last sync finished. This introduces a bug when
a Gradle file is modified after a sync has started but before the
sync has ended. For this case, the current implementation of
isSyncNeeded returns NO after the sync has completed. Instead, it
should return YES because the most recent Gradle file changes
weren't accounted for in the latest sync.

This CL resolves the issue by having GradleSyncState keep track
of the start time of the most recent sync and use this as the
reference timestamp for determining whether or not any Gradle
files have been modified.

Test: GradleSyncStateTest#testIsSyncNeeded_*
Change-Id: I08ab8d2ef69a0061e88fcd3d718e90a60afb7229