Make sure module is not null before calling getModuleIcon
New icons are used after ag/2816987, but for AndroidModuleNode.java, it was
assumed that the module will always not be null but that is not the case.

This CL checks that module is not null before trying to get its icon. I also
confirmed that this was the only place where this was missing.

Bug: 67838273

Test: N/A. Already check manually that no exception is caused when removing
a module

Change-Id: Ib838c5214a2e489f186c2d938d25ea0d5626f07c