Adds a 0-allocation notice for the user
When the user selects a range in the timeline
that has no allocations or deallocations, they
get a near-empty, unhelpful L3 view. This CL
adds and overlay that lets the user know to
select a range with memory activity (or in the
case of pre-O, that there are no allocations
in the currently selected capture).

Change-Id: I55e21d6aad76abb0a3ac4ee035db2a8075b8f6c0
Fixes: 68157666
Test: Updated test to verify info is shown based on selection