Rewriting NavigationSchema
Now it's simpler: it just has two maps:
- PsiClass -> tag
- tag -> DestinationType

This covers all the common cases easily, and the code is simpler.

Now it allows multiple classes to map to the same tag, which wasn't
supported previously but is a valid use case.

It also uses the type of the destination (that is, the type parameter
to Navigator) to decide what the type of a NavigatorSubclass is. This
is to support the case where e.g. a custom (or new bundled) navigator
for Fragments doesn't extend FragmentNavigator.

Test: The existing NavigationSchemaTest was using some dummy Navigator
classes that didn't have the right signature. Changed to use the real
library. Also updated tests as needed.

Change-Id: I11ae0f810c4649b0ed5414d74932036adf759553