Remove GradleDependencyManager usages in DependencyManager
Currently, the DependencyManager classes in both palette and
palette2 use GradleDependencyManager to determine whether
a module is missing a dependency for a palette item. In the case of
platte2/DependencyManager, GradleDependencyManager is also used
to add missing dependencies for a given palette item when necessary.
In order to make the designer and designer-tests modules build
system-independent, these classes need to be updated to use the
new project system APIs instead.

This CL resolves the issue by replacing the GradleDependencyManager
usages with calls to the APIs provided by DependencyManagementUtil
(see DependencyManagement.kt). The corresponding tests have been
updated and converted to Kotlin.

Bug:  65171192
Test: palette/DependencyManagerTest
      + palette2/DependencyManagerTest

Change-Id: I6e5d4bd0e8a27aa209af86bbbd82a0e1b66a4519