Show issue explanation when More (Cmd-F1) is invoked on tooltip
The editor has a facility to show the full issue explanation when
you hover over the error in the editor; the tooltip shows a
"More" link with a keyboard shortcut, and when you invoke the
shortcut the tooltip enlarges to show the full issue explanation.

This did not work for third-party lint rules, since these don't
appear in the builtin issue registry, which is how the help provider
was mapping the issue url back to an issue explanation.

This CL fixes that by looking up the dynamically registered
inspections and finding the issue there.

Test: Unit test included
Change-Id: I416a29ad5586a11c1329ffc97ac848ad84a93101