The event-driven indexing suspender service.
This CL factually embraces all the goals to achieve full
indexing control and reduce associated overhead to an absolute zero.

A bonus is that the newly-introduced state machine which controls
indexing is entirely event-driven, and this means it is fully
decoupled from the build/sync code - it even does not expose any
public methods except for the constructor.

This is important from the code organisation point of view because
build/sync code is full of async calls and various combinations of
possible control flows, which makes it more error-prone to let indexing
control logic depend on that, especially given the dependency on
threading.

Therefore, the only expectation IndexingSuspender has of the rest of
the codebase is that the events will actually signify
what they mean. But whenever that's not the case (or their order is
incorrect), it means there is a much bigger problem than indexing
overhead.

Finally, from the testing point of view, subscription to project
events means that the entire UI test suite will test IndexingSuspender
to some or another extent, which makes regressions much less likely.

Bug: b/67902368, b/67494111, b/67903352, b/67857160
Test: Covered by existing UI test suite. Integration tests TBD once
b/67738862 is complete.

Change-Id: Iee94cb168a7f0d61806369ae7b26828aea86e2cb