Be more aggressive with TooltipComponent removal logic
The TooltipComponent used to get removed only when its parent
was removed from a displayable hierarchy in Swing. This should
have worked, but in practice this can cause a crash internal to
Swing - it seems the tooltip is removed at an unexpected time
and it's possible to get an out-of-bounds exception.

Now, we more aggressively remove the tooltip anytime the user's mouse
exits the tooltip area, and so far it seems like this case is no longer
triggerable.

Change-Id: I2e13df91f1eb1fadc897d31c34de8f90628a02da
Fixes: 68663587
Test: Updated