Add CPU info panel for invalid selection
This CL adds an info panel for when the user selects
an invalid range in the CPU timeline. The panel
contains information for the user to select a valid
range or start a CPU capture.

Also included are some refactoring to get this all
working, such as splitting up certain aspects, as
well as splitting up the CPU capture state machine
with the view.

Change-Id: I583b8c62f2cd8686f52fccba8d725739696dfb34
Fixes: 67941519
Test: updated existing tests