Allow additional process listeners for emulators
In I81235a84b31789b2eb87bdb8c96209159ea021cd, the line adding 'extra'
listeners to the process handler was removed.  The intent of it was to
remove the one place it was being done, in favor of an implementation in
EmulatorProcessHandler.  Partially revert that, allowing the extra
listeners to be added.  Instead remove the listener that was previously
the only user of that feature.

For integration tests, we need to be able to monitor the emulator
process without a UI.  Add a method to AvdManagerConnection that can be
overridden to add listeners consistent with the test logic.

Test: Verified manually that the listener removed here is never added
(thus is safe to remove).  Use of extra listeners in integration tests
coming soon.

Change-Id: Icd77660725ef410900ef5e3592ead14386a9da7a