"Make Project" action now only builds app modules and modules without dependents.
With this CL we avoid iterating through each module in
the project to obtain the task to run to built it. Instead,
we only explicityly build:

a. App modules
b. Non-app modules that no other module depends on

This way, both the IDE and Gradle perform less work:

a. IDE no longer gets the Gradle task to invoke for
   every module.
b. Gradle will still build all the modules in the
   project, but it won't duplicate work for modules
   that are in the dependency graph of app modules

Bug: 68723121
Test: JUnit tests.
Change-Id: If3bde5177218f857b9aa69fad0e436de079df08c