Get SyncMessage type from SyncIssue.
When creating SyncMessage from a SyncIssue,
we sometimes always use MessageType.ERROR, but
the type should come form SyncMessage.getSeverity.

Because ERROR type always cause sync failure, but if
the SyncIssue is SEVERITY_WARNING, it should not fail sync.

This change added a method to get MessageType from
SyncIssue, and call this method when creating SyncMessage
for a SyncIssue.

Bug: 68877171
Test: MessageTypeTest
Change-Id: I57d1287bcd5d1ce8f9199bd43583f05056f696fa