Cause quickfix that adds Google Maven repository to apply to project first.
When there is a module that has a missing dependency, we show a quickfix to
add Google Maven repository if that dependency can be solved from there. The
quickfix first tries to add the repository to module's build file. It should
try to add it first to project build file, in that way all modules that
require it will be fixed in a single edit, instead of applying it to each
one.

This CL causes the quickfix to try to add Google Maven repository first to
project build file.

Bug: 68657672

Test: Modified testReportWithAppCompat to confirm that the repository will
be added to the project build file.

Change-Id: I1e3c380d3391e5c9d66266f6c67bdfdd04f6e1be