Variables panel for the PSD
This adds a new variables panel for the PSD. This panel displays a tree
table of variables, where those variables are displayed by the module
they are defined in.
The values of the variables can be edited in the table using a text box
that offers suggestions of existing variables.

Bug: 68139737
Test: UI tests to be added once agreed on the final UI
Change-Id: I3c4e834544b861e4ebcdabf61db9fb942c38dadd