Handle failed sync when building project.
This CL fixes a bug where if sync fails, the
number of "leaf" modules will be zero, something
that GradleTaskFinder does not expect (it expects
that at there is at least 1 module).

To fix this bug, GradleTaskFinder returns
"assemble" as the task to invoke if the number
of modules is zero. In addition,
MakeGradleProjectAction now only invokes
ProjectStructure#getLeafModules only if the
previous sync was successful.

Bug: 68723121
Test: JUnit test.
Change-Id: I25c4bad2858c40354c3c551126e7a9270980cc93