Refactor utility methods out of IdeProfilerComponents
We have a lot of code that's passing around IdeProfilerComponents as
a glorified bag of utility methods; however, as the class grows, this
feels more and more like an overwielded hammer.

The purpose of IdeProfilerComponents is to provide components, so
those utility methods are cleaned up and encapsulted into their own
custom component class.

This will allow us, in followup CLs, to only pass just the components we
need to a class at construction time, instead of the ever-growing
IdeProfilerCompenets instead.

Test: N/A, refactoring only.
Change-Id: I7ab4c914e1214373b990adea68ac1f1a7363157d