Add method that adds dependency without sync.
GradleDependencyManager offers APIs for adding dependencies to the build
script. This CL adds a version of addDependencies that doesn't
automatically trigger a gradle sync, called addDependenciesWithoutSync.
Usage of this new API is added to GradleModuleSystem, and a gradle sync
is manually triggered after adding the dependency.

Test: Added tests for checking sync does / does not trigger. Manually
checked that invoking a write on the build script without
sync will cause a "request for sync" bar to appear.  Existing tests
pass.

Change-Id: I4b3d3efdd9c0836c9c265a1211adaadcacb460f5