Zoom to fit when switching to design or blueprint only.
This fixes UI tests that had been relying on the design surface to be
zoomed, which was happening in previous tests. Now every test that
chooses to select either design or blueprint view only will
automatically zoom to fit at the same time, which makes it more reliable
to perform tasks such as creating baseline constraints.

Bug: 69316005
Test: ConstraintLayoutTest
Change-Id: I0fca27c48f8a283430bf1b792c73fd00feba0eea