Move Groovy specific code to GroovyDslWriter.
This change moves the groovy specific code for the GradleDslElement
and GradleDslLiteral into the GradleDslWriter. These GradleDsl* classes
now call into methods on the writer to write themselves to the file.
Note: this requires exposing more functionality of the
GradleDslElements, this should be fine since these should only be
accessed via the build model abstractions.

This also changes GroovyPsiElement -> PsiElement in most
GradleDslElement classes.

Bug: 67846255
Test: Existing
Change-Id: Ied893f7ff5d4d475f99041573c9c052f9fe245ac