Refactor CanvasResizeInteraction
The DesignSurface.getCurrentSceneView() may not be
the same one after CanvasResizeInteraction is created.
So we should avoid use DesignSurface.getcurrentSceneView()
in CanvasResizeInteraction and put the target SceneView
into the constructor of it.

Also put Configuration into the constructor of
CanvasResizeInteraction. It is created for updating the
Configuration, so its target Configuration should not be
changed as well.

Test: Passed CanvasResizeTest.
Change-Id: I0be42e1b22cce7f6544de6bda869ce2a824a0672