Remove Session dependency from ProfilerTable and co
The tables that manage profiler Devices and Processes are
now indexed by a device id. The id is generated from SHA based on
the device's serial and boot_id, which are the same two properties
the Session object is created from before. As a result, Device and
Process related queries are no longer dependent on Session.

The rest of the profiler code now passes around a Session object
that only has its device_id set.

Bug: 67508745
Bug: 67508221
Test: Updated unit tests

Change-Id: If49125a6558da53db001d043c11ecb4b20849052