New Payload class in HttpData
Previously, HttpData's API required an external caller to set its
payload file, which is kind of awkward - basically, another class
(currently, ConnectionDetailsView) would own the logic to ask
"Hey, do you have a payload file yet? No? Here, I'll create one
and hand it to you. Yeah it's a good payload file.... trust me!"

As an alternate approach, we now introduce the new HttpData.Payload
class, which is initialized with an immutable payload ID, and it
also owns the file creation logic. A payload's file can be
requested on demand (and once requested, will be cached).

We also introduce an additional API to fetch just the bytes of the
payload, without wrapping it in a file, since there was one place
in the code that only cared about payload bytes.

Finally, note that there's a slight change of behavior from before:
previously, ConnectionDetailsView would try to load payloads
immediately once a request was selected. Now, that logic happens
lazily, when needed. The user shouldn't reeally notice anything,
but a benefit is before setSelectedData could fail if an
IOException was thrown. Now, IOExceptions just result in a
payload not being shown.

Test: Updated
Change-Id: Iecfb89d8eb5aa8ba4dc8d82fafb6153a0cc41333