Introduce BeginSession+EndSession in the profiler update loop
- The two two calls wraps the previous startProfiling/stopProfiling
calls. This change migrates the session concept to be a "profiling"
session instead of a "device" session.

- Added a barebone Profiler_Sessions table to support studio-side
caching and retrieval of sessions.

- Added a special case in the update loop to retrieve the last
available session after a device/process is dead. This handles the case
where data needs to remain in the UI even after a session has ended.

- Removed the session id lookup logic within the DatastoreService. The
id is now retrieved directly from perfd and properly cached in the
database/StudioProfilers. In a follow-up CL, the redundant pid columns
in the data tables can be removed, and all grpc calls will switch to
using the session id instead of the entire Session object.

Bug: 67509046
Bug: 67509029
Bug: 67508808
Bug: 67508775
Bug: 67508745
Test: added/updated unit test
Change-Id: I529c3d163f288ef3a6792f299d52685851211574