Remain smart during build.
This CL amends the indexing suspension mechanism to lift the sentinel dumb
mode while build is in progress. This will allow the users to work with the
IDE full functionality and remain productive while the build is running. We
have seen a fraction of reports saying that fully-fledged code navigation
should be retained during build, so this CL is to address that.

The build remains under a batch update, which prevents the after-sync
indexing caused by model changes from kicking in at the start of the build,
but allows the IDE to remain smart at the same time.

Sync remains both under a batch update and a sentinel dumb mode, which
helps retain no concurrent indexing guarantee where it's most important.

Also, since source generation is treated as a regular build by indexing
suspender, this CL should also alleviate the concerns associated with
the UI impact caused by prolonged dumb mode (for example, see b/64980949).

Test: Integration tests for the suspender are coming in ag/3227103. Verified
all the principal edge cases manually for now (new project, a recent project,
a project which hasn't been opened for a while).
Fixes: b/69455108
Change-Id: I62cd790645be4717ca1dbf03d9ef13769a8cd1c8