Add AVD Wizard support for Virtual Scene.
Virtual Scene Camera is currently being added to the emulator and will be the
new default camera in emulator versions where it is supported.  This change adds
a check for the VirtualScene advanced property and, when available, sets it as
the default back camera in the emulator.  When it is not available, the option
is hidden in the avd wizard.

Test: Manually tested creating and editing with and without virtualScene feature
set in advancedFeatures.ini

Change-Id: I728186a6b1997148bb76730f0a090c81c837b116