Now GradleSyncListener#syncStarted identifies when a sync is skipped.
A sync is skipped when reopening a project and the IDE detects that
build files have not change. In this case, Gradle models are loaded
from a disk cache, instead of performing a full Gradle Sync.

This notificiation is needed to skip entering dumb mode when a sync
is skipped.

Bug: 69665871
Test: JUnit test
Change-Id: I23f28a39d4b1d7cadcf3a43b4f11d8acb9d8a11e