Add ClearResourceCacheAfterFirstBuild
DelayedInitialization is responsible for clearing the
resource cache after source generation is complete and for waiting
to initialize the designer and preview
tabs of the layout editor until after the resource cache has been
cleared. Because its implementation is heavily dependent on Gradle,
DelayedInitialization is currently disabled for Blaze projects.

This CL resolves the issue by introducing
SyncUtil#listenUntilNextSuccessfulSync, a build system-independent
API that allows callers to listen for sync results until after
a successful sync. In order to provide a way
for NlEditorPanel and NlPreviewForm to wait until the project's sync
state is clean before initializing (another side effect of
DelayedInitialization), this CL also introduces
listenForNextSyncResult().

This CL replaces DelayedInitialization with
ClearResourceCacheAfterFirstBuild, which uses
listenUntilNextSuccessfulSync() to clear the resource cache
after the first successful sync with source generation (if necessary)
and provides an API for executing callbacks after the cache
has been validated.

Bug:  65213155
Test: ProjectSystemSyncUtilTest
      + SyncUtilTest
      + ClearResourceCacheAfterFirstBuildTest

Change-Id: Ia6b3d31e934bc324710aab5d1a8f8229a085177e