Remove the last reference to groovy classes from JavaVersionDslElement.
This moves the groovy specific code from
JavaVersionDslElement#getPsiElement into GroovyDslWriter. Due to the
way the Psi tree is constructed it is sometimes the case that the
underlying PsiElement is a GrCommandArgumentList, this means the
element only contains the value literal (e.g only contain the "1.6"
element from "sourceCompatibility 1.6"). When working with this
type of element we need to use the parent, otherwise any operation
won't contain the "sourceCompatibilty" section.

Bug: 67846255
Test: Existing
Change-Id: I600b946533d37bdaf27e146ca12c64938becf218