Android Studio can't load info of older Add-ons
Recent SDK add-ons are described by a package.xml file. For old add-ons
that file is automatically generated based on old add-on format files.
When converting a local SDK add-on, we scan source.properties for
"Addon.NameDisplay" and then for "Addon.Name", and if no value are
found, we assume the "Unknown" value.
Added CL to extend search to manifest.ini too, and
to re-generate package.xml for package.xml with "Unknown" display names.
This CL is part of a topic, and fixes a broken test.

Test: Updated LegacyLocalRepoTest and fix AvdManagerConnectionTest
Change-Id: Icd9b23aef8f2bc1dfb1a37dc76e153bfd4941799