Add GradlePropertyModel as part of the GradleBuildModel.
This change adds a new model that is obtainable from ExtModel
to represent each of the property that have been defined on
the ExtraPropertiesExtension and any Groovy variables thy depend
on.

It extends GradleDslExpression with a way of getting the unresolved
value. e.g "${valueRef}". Providing overrides for all subclasses.

Also creates a GradlePropertyModel interface and provides the first
version of an implementation GradlePropertyModelImpl.

Bug: 67519210
Test: Added unit tests
Change-Id: Ib2cf2833519511e52a3756d11607550f373b754e