[AVD Mgr] Update the skin files if newer versions are available
When creating an AVD and when launching an AVD, check if the
device skin files in <SDK>/skins are up-to-date relative to
the skins files in <STUDIO>/device-art-resources.

When creating an AVD, all skins are checked. When launching an
AVD, only the skins for that AVD are checked.

The "check" is on the timestamp of the 'layout' file. If that file
is missing or old, the device's directory under <SDK>/skins/ is
deleted and re-created using the files in <STUDIO>/device-art-resources.

Bug: 70279040

Test: Added AvdWizardUtilsTest.testEnsureSkinsAreCurrent()

Change-Id: Id9742c8d02c9544ab0e40023d93052b018938e28