Differentiate user code in simpleperf captures
For simpleperf captures, particularly for native functions, we can check
if the function symbol was obtained from a library file belonging to the
app's data directory. If it was, we consider that method as being a user
method. This approach is way more accurate than trying to guess vendor
and platform methods, and assuming the rest of the methods are user's.

Test: Unit tests
Bug: 68014311
Change-Id: I2a143c08bfb9bf80be04a02ffb6dd701ba545acc