Replace gradle APIs with project system APIs.
Replace GradleSyncListener API calls in waitForGradleProjectSyncToFinish
and use project system's sync manager APIs instead.

Note that this CL keeps the existing behaviour of
waitForGradleProjectSyncToFinish and adds documentation for the race
condition it introduces where the method will not return if it is called
after the sync that it's supposed to wait for completes (b/70937990).
This CL does not fix the bug.

Test: Existing tests pass. Manually ran ui-tests.

Change-Id: I571567aa56b1159ba5bdee364a3ad9a87f9fe664