Adds support for setting basic variables and properties.
This change adds support for setting the values of properties
and variables that already exist. Properties that are maps or
lists are not currently supported.

getResolvedVariables is overwritten to ensure depedencies of
newly set variables are correctly computed. We change
GradlePropertiesDslElement to ensure that changes to variables
are applied. And make a small change to GroovyDslUtil to allow
the setting of non-raw strings.

Bug: 67519210
Test: Unit tests added.
Change-Id: I9c1a134ffe2af85c0bb5c208bb6ebe980457eb80