Remove the xlarge variation on the orientation dropdown.
The orientation dropdown has options for creating variations
of the current layout. There are 2 quick selections: landscape
and xLarge. This CL will remove the xLarge selection and replace
it with a Tablet selection.
The new variation will not be available if there is already an
existing variation with minimum width >= 600dp.
If selected it will create a layout variation of 600dp min width.

Bug: b/68153625
Test: Added unit test for OrientationMenuAction.
Change-Id: I22a1c42f76a4c0c3133faaba840b07f1b1d54bd7