Fix property model references for maps and lists.
This change ensures that references are handled correctly for
list and map types in the GradlePropertyModel. It also changes
the way other references (int/str/bool) are handled to ensure
consistency across all of the types.

To enable this we make several changes, first we generalize
GradleStringInjection to store a GradleDslElement instead of
a GradleDslExpression to allow it to work with lists and maps.
We also renamed it to better reflect its purpose.

A new ValueType has also been added to indicate that a property
is a REFERENCE, this type contians a value equals to the name
of the refered to property.

Bug: 67519210
Test: Unit tests added and updated
Change-Id: I9d152d14b3d695f8812ae8d96caf1162653db59b