Replace GradleDependencyManager usages in AppBarConfigurationDialog
Currently, AppBarConfigurationDialog uses GradleDependencyManager
to add a dependency and sync if necessary. In the interest of making
this feature available in ASWB, this CL replaces the
GradleDependencyManager usages with calls to build system-independnet
APIs in DependencyManagementUtil and ProjectSystemSyncManager.

Since automatically adding dependencies for Blaze projects is still
broken, this feature will still not work correctly in ASWB.

This CL also resolves a bug where if the sync after adding a dependency
fails, then the configuration dialog will never stop the loading
animation and report the issue to the user. A test case has been added
to AppBarConfigurationDialogTest to cover this scenario.

Test: Updates to AppBarConfigurationDialogTest
Bug: 70895368

Change-Id: I2382cefc552bbfdc97772742dbb6bc555fa77d3a