Do not wrap checked icon with PoppedIcon in designer.
The IJ ActionMenuItem would wrap the selected icon from a
ToggleAction in a PoppedIcon which raises the icon to make it
look like a button.
This works great for a toolbar button but not for a menuitem in
a popup menu.

Changed the code in TargetMenuAction to not use a ToggleAction to
avoid the raised icon in the popup menu.
The menu changed is: "Automatically Pick Best" under API selector.

Bug: b/69387332
Test: NA. Visual check for now.
Change-Id: I38e98476631e3c06e8085748b74316a5c5229492