Fix plugin updater for composite builds.
Current plugin version updater only updates the
build files and gradle wrapper of current project.
This change will search for build files and gradle
wrapper files in included projects as well, and update
the version for all of them.

Add parameter to BuildFileProcessor, so that caller
can specify whether apply processor on composite builds
or not.

Bug: 70985245
Test: BuildFileProcessorTest.
Change-Id: Ie449ffe277d88f363b3da30fd530ac6fd30419e8