Fix InstantAppRunTest.runFromCmdLine
The test was failing because the test depended on SimpleInstantApp.
SimpleInstantApp was replaced with TopekaInstantApp. The test was
updated to use TopekaInstantApp.

TopekaInstantApp was also updated so that it would not have
dependency conflicts. The :topekabase module used the databinding
feature, which pulled in development versions of the databinding
libraries. These dev versions then pulled in newer versions of the
support libraries, but not all support libraries were updated to
newer versions. This caused a version conflict.

Bug: 70732154
Test: InstantAppRunTest
Change-Id: Ia40365e6d93cc776017582ed4b907c26a5eb079c