FlavorsExecutionTest.runBuildFlavors too slow
This test timed out on buildbot constantly. The cause is that
the layout editor can slow down the robot's menu lookups, since
the editor itself has many menus that the robot has to find and
search. The fix was to close the layout editor after we were
finished with them.

We also scroll the launcher activity checkbox into view on the
Configure Activity dialog. This prevents an error where the
window is not large enough to show the launcher activity checkbox.

Test: FlavorsExecutionTest.runBuildFlavors
Bug: 70729732
Change-Id: I223be6ed21d11a5084686c0ef8861a70ece326c7