Fix flakiness in EspressoRecorderTest.addDependencyOnFly
EspressoRecorderTest.addDependencyOnFly had two sources of
flakiness:

1. The popup menu that appears from following the menu path
"Run" > "Run..." should show an option to run the recorded
Espresso test. This option does not always appear on UI tests,
but it appears consistently when testing manually.

2. The emulator name does not present itself as "device under test"
all the time.

The fix to the above two issues are:

1. Generate a mouse click event within the editor of the newly
opened test file for the recorded test. This is a workaround.
The underlying cause to the original error is still unknown.
See bug 71516507.

2. Kill the emulator after the test is recorded.

Bug: 37150124
Test: EspressoRecorderTest
Change-Id: I155e13adb9179a7a6bf236f6bcb830fdf870cef0