Fix unexpected record type when parsing a simpleperf trace
Handle the META_INFO record data when parsing the simpleperf trace to
avoid throwing a warning when reading such field.

Please note this CL's goal is just to get rid of the warning. For now,
we're just storing the data in a field of SimpleperfTraceParser. We have
plans to use this data in the future when handling multiple event types.

Fixes: 71548054
Test: Just adding a private field to store the content of the data type.
      Automated tests should be added when we add the feature to handle
      these data (e.g. threads shedule in/out).

Change-Id: I081d2ef3915367a31b81b61f948ff9ebb5d41a7f