Move ClassMaps to be a module level service
ClassMaps provides a mapping from a tag names used for elements inside
an XML file to their corresponding PsiClass instances. This was
unnecessarily tied up as part of the AndroidFacet only because it needs
to be cleared when the roots change (libraries are added/removed)

This CL replaces migrates all uses of ClassMaps to a simpler
ModuleService called TagToNameMapper. The implementation takes care of
clearing its cache by listening to facet configuration changes. This
removes the circular dependency between the AndroidFacet and ClassMaps.

Bug: 71553467
Test: existing tests (ClassMapsTest + dom test cases)
Change-Id: Ie48c058777badb4ecb40f648180c4761413dd1b4