Simplify ModuleAvds and rename it to AvdManagerUtils
The main intention behind this CL was to remove the dependency from
AndroidFacet to ModuleAvds. It turns out that ModuleAvds had a cache of
AvdManager that was cleared when the facet was disposed. However, this
cache was entirely unnecessary since AvdManager.getInstance itself
internally had a cache.

The methods inside ModuleAvd that were used in a single place were moved
directly into those classes, leaving it with just two methods, which are
now in AvdManagerUtils.

Bug: 71553467
Test: no functionality change, existing tests
Change-Id: I9329b0f92404c682eb9ef33f3be87184e9586b95