Update HTreeChart template param to be node
... instead of the node's wrapped data.

So, HTreeChart<StringNode>, instead of HTreeChart<String>

Conceptually, an HTreeChart is a chart that wraps a bunch
of nodes. This is not an implementation detail - the
classes' API exposes its nodes, e.g. HTreeChart#getNodeAt.

Moreover, there is no API call which exposes the underlying
node data. That is, there's no way to get a String out of
an HTreeChart<String>.

Finally, there's no need to allow a single HTreeChart
to accept multiple different node types. For example, this
just doesn't happen in our codebase:

 HTreeChart<String> chart = ...
 chart.setHTree(new Utf8Node("...");
 // ... later ...
 chart.setHTree(new Utf16Node("...");

Even if we wanted to do that, we could just create a new chart.
Or create a common base class, e.g. UnicodeNode.

This change touches a lot of files, but in quite a few places
simplifies the parameter or return type (while some class
definitions got a bit more complex in the trade-off)

Test: Updated
Change-Id: I49ca829e2b477e798d842ef215b76a6d1564769a